Local files with external webserver
In this branch, the game is served using a webserver configured to serve python scripts. The wavurl() and imgurl() functions return the relative pathnames to the wav and image files, which are configured to be served directly by the webserver. The wav and image files have been named the same, so that the first value in each record in the "frogs" structure is the base name of the image and wav file without the extension. So, for instance, in this case the image of the American Bullfrog can be accessed at "/Frogwatch/AmericanBullfrog.jpg" and the audio file can be accessed at "/Frogwatch/AmericanBullfrog.wav".